% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{as_interactive_plot}
\alias{as_interactive_plot}
\title{Convert a benchmarking ggplot to an interactive plotly}
\usage{
as_interactive_plot(p, height = NULL)
}
\arguments{
\item{p}{A \code{ggplot} from \code{\link[=static_plot]{static_plot()}} / \code{\link[=static_plot_dyn]{static_plot_dyn()}}.}

\item{height}{Pixel height, or \code{NULL} to let plotly decide.}
}
\value{
A \code{plotly} htmlwidget.
}
\description{
Convert a benchmarking ggplot to an interactive plotly
}
