% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data.R
\name{mod_data_server}
\alias{mod_data_server}
\title{Data Download tab server}
\usage{
mod_data_server(id, base_unit, comparison_units, year_range)
}
\arguments{
\item{id}{Module id.}

\item{base_unit, comparison_units, year_range}{Reactives from
\code{\link[=mod_country_selection_server]{mod_country_selection_server()}}.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Data Download tab server
}
