% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data.R
\name{mod_data_ui}
\alias{mod_data_ui}
\title{Data Download tab UI}
\usage{
mod_data_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\value{
A \code{\link[bslib:nav_panel]{bslib::nav_panel()}}.
}
\description{
Data Download tab UI
}
