% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_overview.R
\name{mod_overview_server}
\alias{mod_overview_server}
\title{Overview tab server}
\usage{
mod_overview_server(
  id,
  primary_iso,
  peer_isos,
  region_codes,
  income_groups,
  year_range
)
}
\arguments{
\item{id}{Module namespace ID.}

\item{primary_iso}{Reactive string -- ISO3 of the focal country.}

\item{peer_isos}{Reactive character vector -- peer ISO3 codes.}

\item{region_codes}{Reactive character vector -- selected region codes.}

\item{income_groups}{Reactive character vector -- selected income groups.}

\item{year_range}{Reactive integer vector of length 2.}
}
\description{
Overview tab server
}
