% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report.R
\name{mod_report_server}
\alias{mod_report_server}
\title{AI Report tab server}
\usage{
mod_report_server(id, base_unit, comparison_units, threshold, year_range)
}
\arguments{
\item{id}{Module id.}

\item{base_unit, comparison_units, threshold, year_range}{Reactives from
\code{\link[=mod_country_selection_server]{mod_country_selection_server()}}.}
}
\value{
Invisibly \code{NULL}.
}
\description{
AI Report tab server
}
