% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report.R
\name{mod_report_ui}
\alias{mod_report_ui}
\title{AI Report tab UI}
\usage{
mod_report_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\value{
A \code{\link[bslib:nav_panel]{bslib::nav_panel()}}.
}
\description{
AI Report tab UI
}
