% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_overview.R
\name{mod_overview_ui}
\alias{mod_overview_ui}
\title{Overview tab UI}
\usage{
mod_overview_ui(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
A \code{\link[bslib:nav-items]{bslib::nav_panel()}} containing the overview dashboard.
}
\description{
Overview tab UI
}
