% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_welcome.R
\name{mod_welcome_ui}
\alias{mod_welcome_ui}
\title{Home tab UI}
\usage{
mod_welcome_ui(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
A \code{\link[bslib:nav-items]{bslib::nav_panel()}} containing the landing page.
}
\description{
Home tab UI
}
