% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cgjrapp.R
\name{run_cgjrapp}
\alias{run_cgjrapp}
\title{Launch the CGJR Shiny application}
\usage{
run_cgjrapp(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}
}
\description{
Starts the Growth and Jobs Institutional Review dashboard.
}
